Add helpful error messages when expected commands are missing
```
$ ./start.sh
minio is up-to-date
Starting rl_coach ... done
robomaker is up-to-date
waiting for containers to start up...
Error: skip showing sagemaker logs because gnome-terminal is not installed.  This is normal if you are on a different OS to Ubuntu.
Error: vncviewer is not present on the PATH.  Make sure you install it and add it to the PATH.
```